input Model 분리
